requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
bson==0.5.10
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
Flask==3.0.0
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
Werkzeug==3.0.0
